% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norman_empodat_batch_data.R
\name{get_norman_empodat_batch_data}
\alias{get_norman_empodat_batch_data}
\title{Batch Retrieve and Paginate Norman Empodat Database}
\usage{
get_norman_empodat_batch_data(
  parameter,
  values,
  saveToDir = NULL,
  dropMemory = FALSE
)
}
\arguments{
\item{parameter}{A character string (e.g., "nsid", "casrn").}

\item{values}{A character vector of values to search for (e.g., c("3380-34-5", "50-00-0")).}

\item{saveToDir}{= when NULL files are not saved, when directory the files are saved as "values.json" file name}

\item{dropMemory}{when true, it clears memory buffer after every loop, data will not be accessible from memory
(json$Data is NULL)}
}
\value{
A consolidated list containing all records retrieved across all
values and pages.
}
\description{
This function extends \code{get_norman_data} by supporting multiple search values
and automatic pagination for the 'empodat' module.
}
