% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_norman_json_files_by_pattern.R
\name{split_norman_json_files_by_pattern}
\alias{split_norman_json_files_by_pattern}
\title{Split JSON Files Based on Partial Key Values (Pattern Matching)}
\usage{
split_norman_json_files_by_pattern(
  input_dir,
  output_dir,
  split_key,
  pattern = NULL,
  data_path = NULL
)
}
\arguments{
\item{input_dir}{Character. Path to the folder containing source JSON files.}

\item{output_dir}{Character. Path to the folder where output files will be saved.}

\item{split_key}{Character. The JSON key to inspect (e.g., "Sampling date").}

\item{pattern}{Character (Regex). A pattern to extract. If NULL, splits by full value.}

\item{data_path}{Character. The element in the JSON containing the records list.}
}
\description{
Split JSON Files Based on Partial Key Values (Pattern Matching)
}
